+++ Sibusten.Philomena.Client/Extensions/PhilomenaImageExtensions.cs
@@ -8,7 +8,7 @@ namespace Sibusten.Philomena.Client.Extensions
 {
     public static class PhilomenaImageExtensions
     {
-        public static async Task DownloadToFile(this IPhilomenaImage image, string filename, CancellationToken cancellationToken = default, IProgress<DownloadProgressInfo>? progress = null)
+        public static async Task DownloadToFile(this IPhilomenaImage image, string filename, CancellationToken cancellationToken = default, IProgress<PhilomenaImageDownloadProgressInfo>? progress = null)
         {
             PhilomenaImageFileDownloader downloader = new PhilomenaImageFileDownloader(image => filename);
             await downloader.Download(image, cancellationToken, progress);