+++ Sibusten.Philomena.Client/Images/IPhilomenaImageSearch.cs
@@ -4,20 +4,14 @@ using System.Threading;
 
 namespace Sibusten.Philomena.Client.Images
 {
-    public record MetadataDownloadProgressInfo
+    public record PhilomenaImageSearchProgressInfo
     {
-        public int Downloaded { get; init; }
-        public int Total { get; init; }
-    }
-
-    public record ImageSearchProgressInfo
-    {
-        public int Processed { get; init; }
-        public int Total { get; init; }
+        public int MetadataDownloaded { get; init; }
+        public int TotalImages { get; init; }
     }
 
     public interface IPhilomenaImageSearch
     {
-        IAsyncEnumerable<IPhilomenaImage> BeginSearch(CancellationToken cancellationToken = default, IProgress<ImageSearchProgressInfo>? searchProgress = null, IProgress<MetadataDownloadProgressInfo>? metadataProgress = null);
+        IAsyncEnumerable<IPhilomenaImage> BeginSearch(CancellationToken cancellationToken = default, IProgress<PhilomenaImageSearchProgressInfo>? searchProgress = null);
     }
 }